+++ CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
-- Add changelog (#11).
+- Add CHANGELOG.md [#11](https://github.com/alkuinvito/chat-client/pull/11).
 
 ### Fixed
-- Enable chat text selection (#10).
+- Enable chat text selection [#10](https://github.com/alkuinvito/chat-client/pull/10).
+- Check username for login [#12](https://github.com/alkuinvito/chat-client/pull/12).
 
 ## [1.0.0] - 2025-08-07